requirements.txt
# RAG Wiki Chatbot Requirements

# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Wikipedia & text processing
wikipedia-api>=0.6.0
mwparserfromhell>=0.6.5
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP & Embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
tokenizers>=0.13.0
tf-keras>=2.15.0  # Required for Keras 3 compatibility with transformers

# Vector DB
faiss-cpu>=1.7.4
# Alternatively: chromadb>=0.4.0

# Retrieval & Reranking
rank-bm25>=0.2.2

# LLM APIs
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0

# Caching
diskcache>=5.6.0

# Evaluation
scikit-learn>=1.3.0
nltk>=3.8.0

# Development
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0


